[feature-60] add authHelper
