+++ src/business/commands/create-user.ts
@@ -3,19 +3,17 @@ import { CredentialsInput } from '../models/inputs/credentials-input';
 import { Credentials } from '../models/results/credentials';
 import { IUserRepository } from '../repositories/i-user-repository';
 import { ICreateUser } from './i-create-user';
-import * as bcrypt from 'bcrypt';
+
 import { flatMap } from 'rxjs/operators';
-import { Errors } from '../utilities/Errors';
+import { Errors } from '../../utilities/Errors';
+import { AuthHelper } from '../../utilities/auth-helper';
 
 export class CreateUser implements ICreateUser {
     constructor(private userRepository: IUserRepository) {
     }
 
     execute(input: CredentialsInput): Observable<Credentials> {
-        const salt = bcrypt.genSaltSync(10);
-        const hash =  bcrypt.hashSync(input.password, salt);
-        const hashedInput = new CredentialsInput({ password: hash, login: input.login });
-
+        const hashedInput = AuthHelper.hashCredentials(input);
         return this.userRepository.byCredentials(input)
                     .pipe(flatMap((user) => iif(() => user !== undefined)
                                 ? throwError(Errors.AlreadyRegisterdError)
